Move `GkrCircuitProof` to air crate
