+++ air/Cargo.toml
@@ -30,6 +30,7 @@ std = ["vm-core/std", "winter-air/std"]
 internals = []
 
 [dependencies]
+thiserror = { version = "1.0", default-features = false }
 vm-core = { package = "miden-core", path = "../core", version = "0.9", default-features = false }
 winter-air = { package = "winter-air", version = "0.9", default-features = false }
 winter-prover = { package = "winter-prover", version = "0.9", default-features = false }
