+++ processor/src/trace/virtual_bus/sum_check/mod.rs
@@ -1,33 +1,11 @@
-use super::univariate::UnivariatePolyCoef;
 use alloc::vec::Vec;
+use miden_air::gkr_proof::RoundProof;
 use vm_core::FieldElement;
 
 mod prover;
 pub use prover::{Error as SumCheckProverError, FinalClaimBuilder, SumCheckProver};
 mod verifier;
 pub use verifier::{CompositionPolyQueryBuilder, Error as SumCheckVerifierError, SumCheckVerifier};
-
-/// A sum-check round proof.
-///
-/// This represents the partial polynomial sent by the Prover during one of the rounds of the
-/// sum-check protocol. The polynomial is in coefficient form and excludes the coefficient for
-/// the linear term as the Verifier can recover it from the other coefficients and the current
-/// (reduced) claim.
-#[derive(Debug, Clone)]
-pub struct RoundProof<E: FieldElement> {
-    pub round_poly_coefs: UnivariatePolyCoef<E>,
-}
-
-/// A sum-check proof.
-///
-/// Composed of the round proofs i.e., the polynomials sent by the Prover at each round as well as
-/// the (claimed) openings of the multi-linear oracles at the evaluation point given by the round
-/// challenges.
-#[derive(Debug, Clone)]
-pub struct Proof<E: FieldElement> {
-    pub openings_claim: FinalOpeningClaim<E>,
-    pub round_proofs: Vec<RoundProof<E>>,
-}
 
 /// Contains the round challenges sent by the Verifier up to some round as well as the current
 /// reduced claim.
@@ -58,26 +36,15 @@ pub fn reduce_claim<E: FieldElement>(
     }
 }
 
-/// Represents an opening claim at an evaluation point against a batch of oracles.
-///
-/// After verifying [`Proof`], the verifier is left with a question on the validity of a final
-/// claim on a number of oracles open to a given set of values at some given point.
-/// This question is answered either using further interaction with the Prover or using
-/// a polynomial commitment opening proof in the compiled protocol.
-#[derive(Clone, Debug)]
-pub struct FinalOpeningClaim<E> {
-    pub eval_point: Vec<E>,
-    pub openings: Vec<E>,
-}
-
 #[cfg(test)]
 mod test {
     use super::{
         prover::{FinalClaimBuilder, SumCheckProver},
         verifier::{CompositionPolyQueryBuilder, SumCheckVerifier},
     };
-    use crate::trace::virtual_bus::multilinear::{CompositionPolynomial, MultiLinearPoly};
+    use crate::trace::virtual_bus::multilinear::CompositionPolynomial;
     use alloc::{borrow::ToOwned, vec::Vec};
+    use miden_air::gkr_proof::{FinalOpeningClaim, MultiLinearPoly};
     use test_utils::rand::rand_vector;
     use vm_core::{crypto::random::RpoRandomCoin, Felt, FieldElement, Word, ONE, ZERO};
 
@@ -170,8 +137,8 @@ mod test {
             &self,
             openings: Vec<Self::Field>,
             evaluation_point: &[Self::Field],
-        ) -> super::FinalOpeningClaim<Self::Field> {
-            super::FinalOpeningClaim {
+        ) -> FinalOpeningClaim<Self::Field> {
+            FinalOpeningClaim {
                 eval_point: evaluation_point.to_owned(),
                 openings,
             }
@@ -184,7 +151,7 @@ mod test {
     impl<E: FieldElement> CompositionPolyQueryBuilder<E> for ProjectionPolyQueryBuilder {
         fn build_query(
             &self,
-            openings_claim: &super::FinalOpeningClaim<E>,
+            openings_claim: &FinalOpeningClaim<E>,
             _evaluation_point: &[E],
         ) -> Vec<E> {
             openings_claim.openings.to_vec()
